requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.8.0
attrs==23.2.0
beautifulsoup4==4.13.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==43.0.1
cycler==0.12.1
distro==1.9.0
fonttools==4.55.8
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.1.1
groq==0.16.0
grpcio==1.65.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.7
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.8
langchain==0.3.17
langchain-core==0.3.33
langchain-groq==0.2.4
langchain-text-splitters==0.3.5
langsmith==0.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
multidict==6.1.0
narwhals==1.24.2
numpy==2.0.1
orjson==3.10.15
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pip==24.0
plotly==6.0.0
propcache==0.2.1
proto-plus==1.24.0
protobuf==4.25.4
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
streamlit==1.37.0
streamlit-lottie==0.0.5
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.1
wheel==0.43.0
wordcloud==1.9.4
yarl==1.18.3
zstandard==0.23.0
